Implemented dining halls busy alert
add off campus restaurants page, add setup.py.
Setup.py generates constants.py file inside the test folder (so that the tests can actually be run). If for whatever reason there's any other files that need to be generated/edited
before the TA runs it for the first time, add a script for it to setup.py.